removed tutorial folder changed to 0.15 updated the tutorial-part1

+++ tutorial-part1/tutorial-part1.py
@@ -4,11 +4,12 @@ sys.path.append('/Users/alex/Documents/OpenPIV/openpiv-python')
 import openpiv.tools
 import openpiv.process
 import openpiv.scaling
+import numpy as np
 
 frame_a  = openpiv.tools.imread( 'exp1_001_a.bmp' )
 frame_b  = openpiv.tools.imread( 'exp1_001_b.bmp' )
 
-u, v, sig2noise = openpiv.process.extended_search_area_piv( frame_a, frame_b, window_size=24, overlap=12, dt=0.02, search_area_size=64, sig2noise_method='peak2peak' )
+u, v, sig2noise = openpiv.process.extended_search_area_piv( frame_a.astype(np.int32), frame_b.astype(np.int32), window_size=24, overlap=12, dt=0.02, search_area_size=64, sig2noise_method='peak2peak' )
 
 x, y = openpiv.process.get_coordinates( image_size=frame_a.shape, window_size=24, overlap=12 )
